move PollNoteViewModel and PollOption to commons
